refactor: add auth module and env options
