+++ lib/auth.js
@@ -13,7 +13,7 @@ export async function findUserByEmail(email) {
   return rows[0] || null;
 }
 
-export async function addUser(email, passwordHash, groups = []) {
+export async function createUser(email, passwordHash, groups = []) {
   const { rows } = await pool.query(
     'INSERT INTO users (email, password_hash, groups) VALUES ($1, $2, $3) RETURNING id',
     [email, passwordHash, groups]
