+++ lib/db.js
@@ -1,7 +1,23 @@
 import { Pool } from 'pg';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const connectionString = process.env.DATABASE_URL;
+
+const config = connectionString
+  ? { connectionString }
+  : {
+      user: process.env.DB_USER,
+      host: process.env.DB_HOST,
+      database: process.env.DB_NAME,
+      password: process.env.DB_PASS,
+      port: process.env.DB_PORT,
+    };
 
 const pool = new Pool({
-  connectionString: process.env.DATABASE_URL,
+  ...config,
+  ssl: { rejectUnauthorized: false },
 });
 
 export default pool;
