+++ package.json
@@ -12,7 +12,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "cookie": "0.5.0",
-    "pg": "8.11.3"
+    "pg": "8.11.3",
+    "dotenv": "latest"
   },
   "devDependencies": {
     "autoprefixer": "latest",
