+++ pages/api/admin/createGroup.js
@@ -1,4 +1,4 @@
-import { addGroup, getSession, getUsers } from '../../../lib/data.js';
+import { addGroup, getSession, getUsers } from '../../../lib/auth.js';
 import { parse } from 'cookie';
 
 export default async function handler(req, res) {
