+++ pages/api/forgot-password.js
@@ -1,4 +1,4 @@
-import { findUserByEmail, createResetToken } from '../../lib/data.js';
+import { findUserByEmail, createResetToken } from '../../lib/auth.js';
 import crypto from 'crypto';
 
 export default async function handler(req, res) {
