+++ pages/api/register.js
@@ -1,4 +1,4 @@
-import { addUser, findUserByEmail } from '../../lib/data.js';
+import { createUser, findUserByEmail } from '../../lib/auth.js';
 import crypto from 'crypto';
 
 export default async function handler(req, res) {
@@ -16,7 +16,7 @@ export default async function handler(req, res) {
   }
 
   const passwordHash = crypto.createHash('sha256').update(password).digest('hex');
-  await addUser(email, passwordHash, []);
+  await createUser(email, passwordHash, []);
 
   res.status(200).json({ success: true });
 }
